Namespace: rename query_frontier/decomposed/web to call-pattern names (0124)
- query_frontier.py → query_direct.py (single call, no retrieval)
- query_decomposed.py → query_per_fuel.py (3 sub-calls by fuel type)
- query_web.py → query_livesearch.py (live Tavily search injection)
- Rename matching test files (test_query_direct/per_fuel/livesearch)
- Update all imports in worker.py, verify.py, harness.py, scripts/
- Update patch paths in tests
- Module docstrings updated to state call pattern
