+++ src/aedist/query_direct.py
@@ -1,18 +1,18 @@
-"""Frontier deep-research benchmark: comprehensive Vietnam thermal sector report.
+"""Single direct LLM call, no retrieval context: comprehensive Vietnam thermal sector report.
 
 Sends a maximally-optimized comprehensive prompt to frontier reasoning models
 and captures full responses for qualitative evaluation.  Uses high max_tokens
 and temperature=0 for deterministic, exhaustive output.
 
 Usage:
-    uv run python -m aedist.query_frontier \
+    uv run python -m aedist.query_direct \
         --prompt prompts/prompt_complete.txt \
         --models models_frontier.yaml \
         --output outputs/frontier/ \
         --budget-usd 20
 
     # Single model:
-    uv run python -m aedist.query_frontier \
+    uv run python -m aedist.query_direct \
         --prompt prompts/prompt_complete.txt \
         --models models_frontier.yaml \
         --output outputs/frontier/ \
